Ten queries averaged an interval and asked for a float
    error occurred while decoding column "avg_wait_minutes": mismatched
    types; Rust type `Option<f64>` (as SQL type `FLOAT8`) is not compatible
    with SQL type `NUMERIC`

PostgreSQL 14 changed `EXTRACT(EPOCH FROM interval)` from `double precision`
to `numeric`, and `AVG(numeric)` is `numeric`. Anything written against an
older server and decoded into `f64` now fails — this database is on 17.

Ten aggregates had no cast. The queue board that reported it is one; the
others are a visitor log, equipment downtime and uptime, case ageing, an ICU
device-days rate, an HR weekly total, and average length of stay. Every one
sits behind a dashboard somebody opens rather than a path anything exercises
on the way past, which is why they were all still there.

Most of the aggregates in this codebase already carry `::float8` — somebody
met this before and fixed what they were looking at. These are the ones they
were not looking at.

The queries are built as runtime strings, so nothing catches this until a
screen is opened. There is a test now that reads the sources and fails on any
aggregate over `EXTRACT(EPOCH ...)` with no cast, naming the file and the
column.

Checked it bites by taking the cast back off, which took two attempts: the
first search looked for `" ::float8"` and the file says `")::float8"`, so
nothing was removed and the test passed on an unchanged file. A green result
from a check that did not run is the failure mode this test exists to
prevent, so it seemed worth confirming rather than assuming.

+++ medbrains/crates/medbrains-ancillary/src/front_office.rs
@@ -862,7 +862,7 @@ pub async fn visitor_analytics(
         "SELECT vr.ward_id AS department_id, \
          COUNT(*)::bigint AS total_visitors, \
          AVG(EXTRACT(EPOCH FROM (vl.check_out_at - vl.check_in_at)) / 60.0) \
-           FILTER (WHERE vl.check_out_at IS NOT NULL) AS avg_visit_duration_minutes, \
+           FILTER (WHERE vl.check_out_at IS NOT NULL)::float8 AS avg_visit_duration_minutes, \
          MODE() WITHIN GROUP (ORDER BY EXTRACT(HOUR FROM vl.check_in_at)::int) AS peak_hour \
          FROM visitor_registrations vr \
          LEFT JOIN visitor_passes vp ON vp.registration_id = vr.id \
@@ -942,7 +942,7 @@ pub async fn queue_metrics(
          EXTRACT(HOUR FROM called_at)::int AS hour_of_day, \
          COUNT(*) FILTER (WHERE queue_status = 'completed')::bigint AS patients_seen, \
          AVG(EXTRACT(EPOCH FROM (called_at - created_at)) / 60.0) \
-           FILTER (WHERE called_at IS NOT NULL) AS avg_wait_minutes, \
+           FILTER (WHERE called_at IS NOT NULL)::float8 AS avg_wait_minutes, \
          MAX(EXTRACT(EPOCH FROM (called_at - created_at)) / 60.0) \
            FILTER (WHERE called_at IS NOT NULL) AS max_wait_minutes \
          FROM opd_queues \
